+++ BinaryTools/BinaryReader.h
@@ -30,7 +30,7 @@ public:
 
     [[nodiscard]] char ReadChar();
     [[nodiscard]] std::string ReadNullTerminatedString();
-    [[nodiscard]] std::string ReadFixedLengthString(int length);
+    [[nodiscard]] std::string ReadFixedLengthString(size_t length);
     [[nodiscard]] char NextChar();
 
     [[nodiscard]] float ReadFloat();
@@ -38,10 +38,10 @@ public:
 
     void ReadToMemory(void* destination, size_t size);
 
-    void SeekBeg(int absoluteOffset);
-    void SeekCur(int relativeOffset);
-    void Skip(int bytesToSkip);
-    size_t Align(int alignmentValue = 2048);
+    void SeekBeg(size_t absoluteOffset);
+    void SeekCur(size_t relativeOffset);
+    void Skip(size_t bytesToSkip);
+    size_t Align(size_t alignmentValue = 2048);
 
     size_t Position() const;
